Extract Subscriber and MpuRequestContext to separate classes
